Refactored the complete Backend code for better readability
