+++ server/.prettierrc
@@ -1,7 +1,7 @@
 {
-    "singleQuote": false,
-    "bracketSpacing": true,
-    "tabWidth": 2,
-    "semi": true,
-    "trailingComma": "es5"
+  "singleQuote": false,
+  "bracketSpacing": true,
+  "tabWidth": 2,
+  "semi": true,
+  "trailingComma": "es5"
 }
