+++ server/controllers/user/profileController.js
@@ -15,9 +15,9 @@ const getProfile = async (req, res) => {
   }
 };
 
-const editProfile = async (req, res) => {
+const editProfileByID = async (req, res) => {
   try {
-    const { id } = req.user; // Get 'id' from token
+    const { id } = req.params; // Get 'id' from token
     let updateData = req.body; // Data to update, parsed from request body
 
     // Check if the user exists
@@ -63,6 +63,6 @@ const editProfile = async (req, res) => {
     }
     res.status(500).json({ message: "Server error", error });
   }
-}
+};
 
-module.exports = { getProfile ,editProfile};
+module.exports = { getProfile, editProfileByID };
