+++ server/docker-compose.yml
@@ -4,6 +4,6 @@ services:
   prom-server:
     image: prom/prometheus
     ports:
-    - 9090:9090
+      - 9090:9090
     volumes:
-    - ./prometheus-config.yml:/etc/prometheus/prometheus.yml
+      - ./prometheus-config.yml:/etc/prometheus/prometheus.yml
