+++ server/index.js
@@ -5,7 +5,9 @@ const bodyParser = require("body-parser");
 // const authRoutes = require("./modules/dashboard/authRoute");  ⭕ ***Deprecated***
 const authRouter = require("./routes/auth/auth");
 const profileRouter = require("./routes/user/profile");
-const hospitalroute = require("./modules/hospital/index");
+const hospitalRouter = require("./routes/hospital/hospital");
+const appointmentRouter = require("./routes/appointments/appointment");
+// const hospitalroute = require("./modules/hospital/index");   ⭕ ***Deprecated***
 const client = require("prom-client");
 const { connectDB, corsConfig } = require("./utils");
 require("dotenv").config();
@@ -49,7 +51,10 @@ app.use("/auth", authRouter);
 app.use("/auth", profileRouter);
 
 // Hospital Routes
-app.use("/hospitalapi", hospitalroute); // Hospital-related routes
+app.use("/hospitalapi", hospitalRouter);
+
+// Appointment Routes
+app.use("/hospitalapi", appointmentRouter);
 
 // Start the Server
 app.listen(port, () => {
