+++ server/models/user.js
@@ -1,27 +1,36 @@
 // models/user.js
-const mongoose = require('mongoose');
+const mongoose = require("mongoose");
 
-const userSchema = new mongoose.Schema({
+const userSchema = new mongoose.Schema(
+  {
     name: String,
     email: { type: String, required: true, unique: true },
     password: { type: String, required: true },
     phone: String,
     dob: Date,
-    gender: { type: String, enum: ['Male', 'Female', 'Other'] },
+    gender: { type: String, enum: ["Male", "Female", "Other"] },
     address: {
-        street: String,
-        city: String,
-        state: String,
-        postalCode: String,
+      street: String,
+      city: String,
+      state: String,
+      postalCode: String,
     },
     medicalHistory: [String],
-    appointments: [{
-        hospitalId: { type: mongoose.Schema.Types.ObjectId, ref: 'Hospital' },
+    appointments: [
+      {
+        hospitalId: { type: mongoose.Schema.Types.ObjectId, ref: "Hospital" },
         date: Date,
         reason: String,
-        status: { type: String, enum: ['pending', 'completed', 'canceled'], default: 'pending' }
-    }]
-}, { timestamps: true });
+        status: {
+          type: String,
+          enum: ["pending", "completed", "canceled"],
+          default: "pending",
+        },
+      },
+    ],
+  },
+  { timestamps: true }
+);
 
-const User = mongoose.model('User', userSchema);
+const User = mongoose.model("User", userSchema);
 module.exports = User;
